requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
jinja2>=3.1.2
python-dotenv>=1.0.0
